Emotionのenumを別ファイルに移動
